+++ src/frontend/pages/_app.tsx
@@ -9,6 +9,7 @@ import CurrencyProvider from '../providers/Currency.provider';
 import CartProvider from '../providers/Cart.provider';
 import { ThemeProvider } from 'styled-components';
 import Theme from '../styles/Theme';
+import FrontendTracer from '../utils/telemetry/FrontendTracer';
 import SessionGateway from '../gateways/Session.gateway';
 import { OpenFeatureProvider, OpenFeature } from '@openfeature/react-sdk';
 import { FlagdWebProvider } from '@openfeature/flagd-web-provider';
@@ -68,6 +69,7 @@ if (typeof window !== 'undefined') {
     },
   });
 
+  FrontendTracer();
   if (window.location) {
     const session = SessionGateway.getSession();